requirements.txt
django~=5.0.3
pillow
psycopg2-binary
python-decouple
djangorestframework~=3.15.1
dj-rest-auth
drf-yasg
django-filter~=24.2
djangorestframework-simplejwt
django-ckeditor-5

product~=0.1.1.1
requests~=2.31.0
beautifulsoup4~=4.12.3